+++ src/extension.ts
@@ -4,6 +4,7 @@ import * as vscode from "vscode";
 import { DART_MODE } from "./dartMode";
 import { Pub } from "./pubCmd";
 import { DartFormat } from "./dartFormat";
+import { DartAnalyzer } from "./dartAnalyzer";
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -20,9 +21,11 @@ export function activate(context: vscode.ExtensionContext) : void {
 function init(context: vscode.ExtensionContext) : void {
 	let pub = new Pub();
 	let fmt = new DartFormat();
+	let analyzer = new DartAnalyzer();
 
 	pub.setPubCmd(context);
 	fmt.setDartFmt(context);
+	analyzer.setDartAnalyzer(context);
 }
 
 function setLanguageConfiguration() {